requirements.txt
Django>=4.2.10,<5.0
djangorestframework>=3.14,<3.16
django-cors-headers>=4.0,<5.0
djangorestframework-simplejwt>=5.2,<6.0
psycopg2-binary>=2.9.9,<3.0
gunicorn>=20.1,<22
pillow>=10.3,<11
python-decouple==3.8
dj-database-url>=2.1,<3.0
whitenoise>=6.6,<7.0
requests
razorpay
